requirements/dev_requirements.txt
flake8==3.7.9
PyJWT==1.7.1
pytest==5.3.2
pytest-cov==2.8.1
pytest-flask==0.15.1
pytest-mock==1.10.0
requests==2.22.0
requests-toolbelt==0.9.1